requirements.txt
gunicorn
whitenoise
dj-database-url
psycopg2-binary
Django==5.1.6
kinde_python_sdk==1.2.6
python-dotenv==1.0.1
django-storages
boto3
djangorestframework
django-cors-headers
requests